+++ Test/Resources/CancelationResources/cancelByXml.xml
