+++ Test/Resources/cadenaOriginal.txt
