+++ Test/StampTests.php
